requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-firestore
pydantic
google-adk